GiftHub_serving/requirements.txt
fastapi==0.103.0
uvicorn==0.20.0
pydantic-settings==2.2.1
pandas==2.2.1
mlflow==2.10.0
numpy==1.26.4
lightgbm==4.1.0
cloudpickle==3.0.0